Create 'theme' cmd and 'add' sub cmd.
